Refactor jest tests
